General Clean up for code
